move usages file into constants
